Move run test from base to task model
The wrapper has moved
